+++ chat2db-client/src/constants/common.ts
@@ -32,3 +32,34 @@ export enum OSType {
   MAC = 'Mac',
   RESTS = 'rests',
 }
+
+export enum OperationType {
+  CONSOLE = 'console',
+  FUNCTION = 'function',
+  PROCEDURE = 'procedure',
+  VIEW = 'view',
+  TRIGGER = 'trigger',
+}
+
+export const operationTypeConfig: {
+  [key in OperationType]: {
+    icon: string
+  };
+} = {
+  [OperationType.CONSOLE]: {
+    icon: '\ue619'
+  },
+  [OperationType.VIEW]: {
+    icon: '\ue70c'
+  },
+  [OperationType.FUNCTION]: {
+    icon: '\ue76a'
+  },
+  [OperationType.PROCEDURE]: {
+    icon: '\ue73c'
+  },
+
+  [OperationType.TRIGGER]: {
+    icon: '\ue64a'
+  }
+}