+++ chat2db-server/chat2db-plugins/chat2db-oracle/src/main/java/ai/chat2db/plugin/oracle/OracleDBManage.java
@@ -1,8 +1,30 @@
 package ai.chat2db.plugin.oracle;
 
+import java.sql.Connection;
+import java.sql.SQLException;
+
 import ai.chat2db.spi.DBManage;
 import ai.chat2db.spi.jdbc.DefaultDBManage;
+import ai.chat2db.spi.sql.Chat2DBContext;
+import ai.chat2db.spi.sql.ConnectInfo;
+import ai.chat2db.spi.sql.SQLExecutor;
+import org.apache.commons.lang3.ObjectUtils;
+import org.apache.commons.lang3.StringUtils;
 
 public class OracleDBManage extends DefaultDBManage implements DBManage {
+
+    @Override
+    public void connectDatabase(Connection connection, String database) {
+        ConnectInfo connectInfo = Chat2DBContext.getConnectInfo();
+        if (ObjectUtils.anyNull(connectInfo) || StringUtils.isEmpty(connectInfo.getSchemaName())) {
+            return;
+        }
+        String schemaName = connectInfo.getSchemaName();
+        try {
+            SQLExecutor.getInstance().execute(connection, "ALTER SESSION SET CURRENT_SCHEMA = \"" + schemaName + "\";");
+        } catch (SQLException e) {
+
+        }
+    }
 
 }