+++ chat2db-server/chat2db-server-domain/chat2db-server-domain-core/src/main/java/ai/chat2db/server/domain/core/impl/DlTemplateServiceImpl.java
@@ -12,7 +12,6 @@ import com.alibaba.druid.sql.SQLUtils;
 import com.alibaba.druid.sql.ast.SQLStatement;
 import com.alibaba.druid.sql.ast.statement.SQLSelectStatement;
 import com.alibaba.druid.sql.parser.ParserException;
-import com.alibaba.druid.sql.parser.SQLParserUtils;
 
 import ai.chat2db.server.domain.api.param.DlCountParam;
 import ai.chat2db.server.domain.api.param.DlExecuteParam;
@@ -31,6 +30,7 @@ import ai.chat2db.spi.model.Header;
 import ai.chat2db.spi.sql.Chat2DBContext;
 import ai.chat2db.spi.sql.SQLExecutor;
 import ai.chat2db.spi.util.JdbcUtils;
+import ai.chat2db.spi.util.SqlUtils;
 import com.google.common.collect.Lists;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.collections4.CollectionUtils;
@@ -58,7 +58,7 @@ public class DlTemplateServiceImpl implements DlTemplateService {
         sqlAnalyseParam.setSql(param.getSql());
         DbType dbType =
             JdbcUtils.parse2DruidDbType(Chat2DBContext.getConnectInfo().getDbType());
-        List<String> sqlList = SQLParserUtils.splitAndRemoveComment(param.getSql(), dbType);
+        List<String> sqlList = SqlUtils.parse(param.getSql(), dbType);
         if (CollectionUtils.isEmpty(sqlList)) {
             throw new BusinessException("dataSource.sqlAnalysisError");
         }
@@ -67,31 +67,33 @@ public class DlTemplateServiceImpl implements DlTemplateService {
         ListResult<ExecuteResult> listResult = ListResult.of(result);
         // 执行sql
         for (String originalSql : sqlList) {
-            String sql = originalSql;
-            int pageNo = 0;
-            int pageSize = 0;
-            String sqlType = SqlTypeEnum.UNKNOWN.getCode();
-
-            // 解析sql分页
-            SQLStatement sqlStatement;
-            try {
-                sqlStatement = SQLUtils.parseSingleStatement(sql, dbType);
-            } catch (ParserException e) {
-                log.warn("解析sql失败:{}", sql, e);
-                ExecuteResult executeResult = ExecuteResult.builder()
-                    .success(Boolean.FALSE)
-                    .originalSql(originalSql)
-                    .sql(sql)
-                    .message(e.getMessage())
-                    .build();
-                result.add(executeResult);
-                continue;
+            ExecuteResult executeResult = executeSQL(originalSql,dbType,param);
+            result.add(executeResult);
+            if (!executeResult.getSuccess()) {
+                listResult.setSuccess(false);
+                listResult.errorCode(executeResult.getDescription());
+                listResult.setErrorMessage(executeResult.getMessage());
             }
+        }
+        return listResult;
+    }
+
+    private ExecuteResult executeSQL(String originalSql,DbType dbType,DlExecuteParam param) {
+        String sql = originalSql;
+        int pageNo = 0;
+        int pageSize = 0;
+        String sqlType = SqlTypeEnum.UNKNOWN.getCode();
+
+        // 解析sql分页
+        SQLStatement sqlStatement;
+        boolean autoLimit = false;
+        try {
+            sqlStatement = SQLUtils.parseSingleStatement(sql, dbType);
             // 是否需要代码帮忙分页
-            boolean autoLimit = false;
             if (sqlStatement instanceof SQLSelectStatement) {
                 //  不是查询全部数据 而且 用户自己没有传分页
-                autoLimit = BooleanUtils.isNotTrue(param.getPageSizeAll()) && SQLUtils.getLimit(sqlStatement, dbType)
+                autoLimit = BooleanUtils.isNotTrue(param.getPageSizeAll()) && SQLUtils.getLimit(sqlStatement,
+                    dbType)
                     == null;
                 if (autoLimit) {
                     pageNo = Optional.ofNullable(param.getPageNo()).orElse(1);
@@ -105,53 +107,51 @@ public class DlTemplateServiceImpl implements DlTemplateService {
                 }
                 sqlType = SqlTypeEnum.SELECT.getCode();
             }
+        } catch (ParserException e) {
+            log.warn("解析sql失败:{}", sql, e);
+        }
 
-            ExecuteResult executeResult = execute(sql);
-            executeResult.setSqlType(sqlType);
-            executeResult.setOriginalSql(originalSql);
-            // 自动分页
-            if (autoLimit) {
-                executeResult.setPageNo(pageNo);
-                executeResult.setPageSize(pageSize);
-                executeResult.setHasNextPage(
-                    CollectionUtils.size(executeResult.getDataList()) >= executeResult.getPageSize());
-            } else {
-                executeResult.setPageNo(1);
-                executeResult.setPageSize(CollectionUtils.size(executeResult.getDataList()));
-                executeResult.setHasNextPage(Boolean.FALSE);
-            }
-            // Splice row numbers
-            List<Header> newHeaderList = new ArrayList<>();
-            newHeaderList.add(Header.builder()
-                .name(I18nUtils.getMessage("sqlResult.rowNumber"))
-                .dataType(DataTypeEnum.CHAT2DB_ROW_NUMBER
-                    .getCode()).build());
-            if (executeResult.getHeaderList() != null) {
-                newHeaderList.addAll(executeResult.getHeaderList());
-            }
-            executeResult.setHeaderList(newHeaderList);
-            if (executeResult.getDataList() != null) {
-                int rowNumberIncrement = 1 + Math.max(pageNo - 1, 0) * pageSize;
-                for (int i = 0; i < executeResult.getDataList().size(); i++) {
-                    List<String> row = executeResult.getDataList().get(i);
-                    List<String> newRow = Lists.newArrayListWithExpectedSize(row.size() + 1);
-                    newRow.add(Integer.toString(i + rowNumberIncrement));
-                    newRow.addAll(row);
-                    executeResult.getDataList().set(i, newRow);
-                }
-            }
-            //  Total number of fuzzy rows
-            executeResult.setFuzzyTotal(calculateFuzzyTotal(pageNo, pageSize, executeResult));
-
-            result.add(executeResult);
-            if (!executeResult.getSuccess()) {
-                listResult.setSuccess(false);
-                listResult.errorCode(executeResult.getDescription());
-                listResult.setErrorMessage(executeResult.getMessage());
+        ExecuteResult executeResult = execute(sql);
+        executeResult.setSqlType(sqlType);
+        executeResult.setOriginalSql(originalSql);
+        // 自动分页
+        if (autoLimit) {
+            executeResult.setPageNo(pageNo);
+            executeResult.setPageSize(pageSize);
+            executeResult.setHasNextPage(
+                CollectionUtils.size(executeResult.getDataList()) >= executeResult.getPageSize());
+        } else {
+            executeResult.setPageNo(1);
+            executeResult.setPageSize(CollectionUtils.size(executeResult.getDataList()));
+            executeResult.setHasNextPage(Boolean.FALSE);
+        }
+        // Splice row numbers
+        List<Header> newHeaderList = new ArrayList<>();
+        newHeaderList.add(Header.builder()
+            .name(I18nUtils.getMessage("sqlResult.rowNumber"))
+            .dataType(DataTypeEnum.CHAT2DB_ROW_NUMBER
+                .getCode()).build());
+        if (executeResult.getHeaderList() != null) {
+            newHeaderList.addAll(executeResult.getHeaderList());
+        }
+        executeResult.setHeaderList(newHeaderList);
+        if (executeResult.getDataList() != null) {
+            int rowNumberIncrement = 1 + Math.max(pageNo - 1, 0) * pageSize;
+            for (int i = 0; i < executeResult.getDataList().size(); i++) {
+                List<String> row = executeResult.getDataList().get(i);
+                List<String> newRow = Lists.newArrayListWithExpectedSize(row.size() + 1);
+                newRow.add(Integer.toString(i + rowNumberIncrement));
+                newRow.addAll(row);
+                executeResult.getDataList().set(i, newRow);
             }
         }
-        return listResult;
+        //  Total number of fuzzy rows
+        executeResult.setFuzzyTotal(calculateFuzzyTotal(pageNo, pageSize, executeResult));
+        return executeResult;
     }
+
+
+
 
     private String calculateFuzzyTotal(int pageNo, int pageSize, ExecuteResult executeResult) {
         int dataSize = CollectionUtils.size(executeResult.getDataList());