+++ chat2db-server/chat2db-spi/src/main/java/ai/chat2db/spi/model/Procedure.java
@@ -37,4 +37,6 @@ public class Procedure {
     private Short procedureType;
 
     private String specificName;
+
+    private String procedureBody;
 }